requirements.txt
arrow==1.2.3
Django==4.2.3
django-cors-headers==4.2.0
djangorestframework==3.14.0
python-dotenv==1.0.0
requests==2.31.0